Added support for offline invalidation, via PPTC, of low cq translations replaced by high cq translations; both on a single run and between runs.
Added invalidation of .cache files in the event of reuse on a different user operating system.

Added .info and .cache files invalidation in case of a failed stream decompression.

Nits.

+++ Ryujinx.HLE/HOS/ProgramLoader.cs
@@ -171,7 +171,7 @@ namespace Ryujinx.HLE.HOS
             }
 
             PtcProfiler.StaticCodeStart = codeStart;
-            PtcProfiler.StaticCodeSize  = codeSize;
+            PtcProfiler.StaticCodeSize  = (ulong)codeSize;
 
             int codePagesCount = codeSize / KMemoryManager.PageSize;
